HandClassifier and initial spec
This spec would to be fleshed out a lot more if it were going to be used for anything more than the ProjectEuler test dataset. In particular, the dataset doesn't really test the hand classification/scoring edge cases nearly enough as it should. This should do reasonably well to cover us for Problem 54.